# pb/filter_message.pb.go
// Code generated by protoc-gen-go. DO NOT EDIT.
// versions:
// 	protoc-gen-go v1.31.0
// 	protoc        v4.24.3
// source: filter_message.proto

package pb

import (
	protoreflect "google.golang.org/protobuf/reflect/protoreflect"
	protoimpl "google.golang.org/protobuf/runtime/protoimpl"
	reflect "reflect"
	sync "sync"
)

const (
	// Verify that this generated code is sufficiently up-to-date.
	_ = protoimpl.EnforceVersion(20 - protoimpl.MinVersion)
	// Verify that runtime/protoimpl is sufficiently up-to-date.
	_ = protoimpl.EnforceVersion(protoimpl.MaxVersion - 20)
)

type FilterMessage struct {
	state         protoimpl.MessageState
	sizeCache     protoimpl.SizeCache
	unknownFields protoimpl.UnknownFields

	MaxPriceUsd float64 `protobuf:"fixed64,1,opt,name=max_price_usd,json=maxPriceUsd,proto3" json:"max_price_usd,omitempty"`
	CpuCores    uint32  `protobuf:"varint,2,opt,name=cpu_cores,json=cpuCores,proto3" json:"cpu_cores,omitempty"`
	MixCpuGhz   float64 `protobuf:"fixed64,3,opt,name=mix_cpu_ghz,json=mixCpuGhz,proto3" json:"mix_cpu_ghz,omitempty"`
	MinRam      *Memory `protobuf:"bytes,4,opt,name=min_ram,json=minRam,proto3" json:"min_ram,omitempty"`
}

func (x *FilterMessage) Reset() {
	*x = FilterMessage{}
	if protoimpl.UnsafeEnabled {
		mi := &file_filter_message_proto_msgTypes[0]
		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
		ms.StoreMessageInfo(mi)
	}
}

func (x *FilterMessage) String() string {
	return protoimpl.X.MessageStringOf(x)
}

func (*FilterMessage) ProtoMessage() {}

func (x *FilterMessage) ProtoReflect() protoreflect.Message {
	mi := &file_filter_message_proto_msgTypes[0]
	if protoimpl.UnsafeEnabled && x != nil {
		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
		if ms.LoadMessageInfo() == nil {
			ms.StoreMessageInfo(mi)
		}
		return ms
	}
	return mi.MessageOf(x)
}

// Deprecated: Use FilterMessage.ProtoReflect.Descriptor instead.
func (*FilterMessage) Descriptor() ([]byte, []int) {
	return file_filter_message_proto_rawDescGZIP(), []int{0}
}

func (x *FilterMessage) GetMaxPriceUsd() float64 {
	if x != nil {
		return x.MaxPriceUsd
	}
	return 0
}

func (x *FilterMessage) GetCpuCores() uint32 {
	if x != nil {
		return x.CpuCores
	}
	return 0
}

func (x *FilterMessage) GetMixCpuGhz() float64 {
	if x != nil {
		return x.MixCpuGhz
	}
	return 0
}

func (x *FilterMessage) GetMinRam() *Memory {
	if x != nil {
		return x.MinRam
	}
	return nil
}

var File_filter_message_proto protoreflect.FileDescriptor

var file_filter_message_proto_rawDesc = []byte{
	0x0a, 0x14, 0x66, 0x69, 0x6c, 0x74, 0x65, 0x72, 0x5f, 0x6d, 0x65, 0x73, 0x73, 0x61, 0x67, 0x65,
	0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x1a, 0x14, 0x6d, 0x65, 0x6d, 0x6f, 0x72, 0x79, 0x5f, 0x6d,
	0x65, 0x73, 0x73, 0x61, 0x67, 0x65, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x22, 0x92, 0x01, 0x0a,
	0x0d, 0x46, 0x69, 0x6c, 0x74, 0x65, 0x72, 0x4d, 0x65, 0x73, 0x73, 0x61, 0x67, 0x65, 0x12, 0x22,
	0x0a, 0x0d, 0x6d, 0x61, 0x78, 0x5f, 0x70, 0x72, 0x69, 0x63, 0x65, 0x5f, 0x75, 0x73, 0x64, 0x18,
	0x01, 0x20, 0x01, 0x28, 0x01, 0x52, 0x0b, 0x6d, 0x61, 0x78, 0x50, 0x72, 0x69, 0x63, 0x65, 0x55,
	0x73, 0x64, 0x12, 0x1b, 0x0a, 0x09, 0x63, 0x70, 0x75, 0x5f, 0x63, 0x6f, 0x72, 0x65, 0x73, 0x18,
	0x02, 0x20, 0x01, 0x28, 0x0d, 0x52, 0x08, 0x63, 0x70, 0x75, 0x43, 0x6f, 0x72, 0x65, 0x73, 0x12,
	0x1e, 0x0a, 0x0b, 0x6d, 0x69, 0x78, 0x5f, 0x63, 0x70, 0x75, 0x5f, 0x67, 0x68, 0x7a, 0x18, 0x03,
	0x20, 0x01, 0x28, 0x01, 0x52, 0x09, 0x6d, 0x69, 0x78, 0x43, 0x70, 0x75, 0x47, 0x68, 0x7a, 0x12,
	0x20, 0x0a, 0x07, 0x6d, 0x69, 0x6e, 0x5f, 0x72, 0x61, 0x6d, 0x18, 0x04, 0x20, 0x01, 0x28, 0x0b,
	0x32, 0x07, 0x2e, 0x4d, 0x65, 0x6d, 0x6f, 0x72, 0x79, 0x52, 0x06, 0x6d, 0x69, 0x6e, 0x52, 0x61,
	0x6d, 0x42, 0x0c, 0x5a, 0x0a, 0x70, 0x63, 0x2d, 0x62, 0x6f, 0x6f, 0x6b, 0x2f, 0x70, 0x62, 0x62,
	0x06, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x33,
}

var (
	file_filter_message_proto_rawDescOnce sync.Once
	file_filter_message_proto_rawDescData = file_filter_message_proto_rawDesc
)

func file_filter_message_proto_rawDescGZIP() []byte {
	file_filter_message_proto_rawDescOnce.Do(func() {
		file_filter_message_proto_rawDescData = protoimpl.X.CompressGZIP(file_filter_message_proto_rawDescData)
	})
	return file_filter_message_proto_rawDescData
}

var file_filter_message_proto_msgTypes = make([]protoimpl.MessageInfo, 1)
var file_filter_message_proto_goTypes = []interface{}{
	(*FilterMessage)(nil), // 0: FilterMessage
	(*Memory)(nil),        // 1: Memory
}
var file_filter_message_proto_depIdxs = []int32{
	1, // 0: FilterMessage.min_ram:type_name -> Memory
	1, // [1:1] is the sub-list for method output_type
	1, // [1:1] is the sub-list for method input_type
	1, // [1:1] is the sub-list for extension type_name
	1, // [1:1] is the sub-list for extension extendee
	0, // [0:1] is the sub-list for field type_name
}

func init() { file_filter_message_proto_init() }
func file_filter_message_proto_init() {
	if File_filter_message_proto != nil {
		return
	}
	file_memory_message_proto_init()
	if !protoimpl.UnsafeEnabled {
		file_filter_message_proto_msgTypes[0].Exporter = func(v interface{}, i int) interface{} {
			switch v := v.(*FilterMessage); i {
			case 0:
				return &v.state
			case 1:
				return &v.sizeCache
			case 2:
				return &v.unknownFields
			default:
				return nil
			}
		}
	}
	type x struct{}
	out := protoimpl.TypeBuilder{
		File: protoimpl.DescBuilder{
			GoPackagePath: reflect.TypeOf(x{}).PkgPath(),
			RawDescriptor: file_filter_message_proto_rawDesc,
			NumEnums:      0,
			NumMessages:   1,
			NumExtensions: 0,
			NumServices:   0,
		},
		GoTypes:           file_filter_message_proto_goTypes,
		DependencyIndexes: file_filter_message_proto_depIdxs,
		MessageInfos:      file_filter_message_proto_msgTypes,
	}.Build()
	File_filter_message_proto = out.File
	file_filter_message_proto_rawDesc = nil
	file_filter_message_proto_goTypes = nil
	file_filter_message_proto_depIdxs = nil
}